back-end/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
sqlalchemy>=1.4.0,<1.5.0
mysql-connector-python>=8.0.26,<8.1.0
pymysql>=1.0.2,<1.1.0
python-dotenv>=0.19.0,<0.20.0
passlib>=1.7.4,<1.8.0
python-jose>=3.3.0,<3.4.0
python-multipart>=0.0.5,<0.0.6
email-validator>=1.1.3,<1.2.0
requests>=2.26.0,<2.27.0
alembic>=1.7.4,<1.8.0
redis>=4.0.0,<4.1.0
sqlmodel>=0.0.8,<0.1.0